+++ RELEASE_NOTES.rst
@@ -52,6 +52,8 @@ Vagga 0.7.0
 * Added ``vagga _clean --volumes`` and ``vagga _clean --unused-volumes``
 * Implemented ``vagga _clean --everything`` (again after 0.2)
 * Upgraded embedded alpine tools (apk-tools 2.6.7, busybox 1.24.2)
+* Vagga does not output environment of a running command any more,
+  use ``VAGGA_DEBUG_CMDENV`` to show
 
 
 Vagga 0.6.1